Starting some cleanup
query_plans.go has gotten very, very large, so it's time for some cleanup.
