+++ plans/query_plans.go
@@ -8,145 +8,9 @@ import (
 	"strings"
 
 	"github.com/go-gorp/gorp"
-	"github.com/nelsam/gorq/dialects"
 	"github.com/nelsam/gorq/filters"
 	"github.com/nelsam/gorq/interfaces"
 )
-
-type fieldColumnMap struct {
-	// addr should be the address (pointer value) of the field within
-	// the struct being used to construct this query.
-	addr interface{}
-
-	// column should be the column that matches the field that addr
-	// points to.
-	column *gorp.ColumnMap
-
-	// quotedTable should be the pre-quoted table string for this
-	// column.
-	quotedTable string
-
-	// quotedColumn should be the pre-quoted column string for this
-	// column.
-	quotedColumn string
-}
-
-type structColumnMap []fieldColumnMap
-
-// LocateColumn takes an interface value (which should be a
-// pointer to one of the fields on the value that is being used as a
-// reference for query construction) and returns the pre-quoted column
-// name that should be used to reference that value in queries.
-func (structMap structColumnMap) LocateColumn(fieldPtr interface{}) (string, error) {
-	fieldMap, err := structMap.fieldMapForPointer(fieldPtr)
-	if err != nil {
-		return "", err
-	}
-	return fieldMap.quotedColumn, nil
-}
-
-// LocateTableAndColumn takes an interface value (which should be a
-// pointer to one of the fields on the value that is being used as a
-// reference for query construction) and returns the pre-quoted
-// table.column name that should be used to reference that value in
-// some types of queries (mostly where statements and select queries).
-func (structMap structColumnMap) LocateTableAndColumn(fieldPtr interface{}) (string, error) {
-	fieldMap, err := structMap.fieldMapForPointer(fieldPtr)
-	if err != nil {
-		return "", err
-	}
-	return fieldMap.quotedTable + "." + fieldMap.quotedColumn, nil
-}
-
-// fieldMapForPointer takes a pointer to a struct field and returns
-// the fieldColumnMap for that struct field.
-func (structMap structColumnMap) fieldMapForPointer(fieldPtr interface{}) (*fieldColumnMap, error) {
-	for _, fieldMap := range structMap {
-		if fieldMap.addr == fieldPtr {
-			if fieldMap.column.Transient {
-				return nil, errors.New("gorp: Cannot run queries against transient columns")
-			}
-			return &fieldMap, nil
-		}
-	}
-	fieldPtrVal := reflect.ValueOf(fieldPtr)
-	addr, value := fieldPtrVal.Pointer(), fieldPtrVal.Elem().Interface()
-	return nil, fmt.Errorf("gorp: Cannot find a field matching the passed in pointer %d (value %v)", addr, value)
-}
-
-type order struct {
-	fieldOrWrapper interface{}
-	direction      string
-}
-
-func (o order) OrderBy(dialect gorp.Dialect, colMap structColumnMap, bindIdx int) (string, []interface{}, error) {
-	var (
-		wrapper      filters.SqlWrapper
-		allFields    []interface{}
-		multiWrapper filters.MultiSqlWrapper
-	)
-	switch t := o.fieldOrWrapper.(type) {
-	case filters.SqlWrapper:
-		wrapper = t
-		allFields = []interface{}{wrapper.ActualValue()}
-	case filters.MultiSqlWrapper:
-		multiWrapper = t
-		allFields = multiWrapper.ActualValues()
-	default:
-		allFields = []interface{}{o.fieldOrWrapper}
-	}
-	// OrderBy needs at least one reference to a column of some sort.
-	fieldFound := false
-	columnsAndFields := make([]string, 0, len(allFields))
-	params := make([]interface{}, 0, len(allFields))
-	for _, field := range allFields {
-		if reflect.TypeOf(field).Kind() == reflect.Ptr {
-			column, err := colMap.LocateTableAndColumn(field)
-			if err != nil {
-				return "", nil, err
-			}
-			columnsAndFields = append(columnsAndFields, column)
-			fieldFound = true
-		} else {
-			columnsAndFields = append(columnsAndFields, dialect.BindVar(bindIdx))
-			params = append(params, field)
-			bindIdx++
-		}
-	}
-	if !fieldFound {
-		return "", nil, errors.New("OrderBy requires a pointer to a struct field or " +
-			"a wrapper with at least one struct field pointer as an actual value.")
-	}
-	var orderStr string
-	if wrapper != nil {
-		orderStr = wrapper.WrapSql(columnsAndFields[0])
-	} else if multiWrapper != nil {
-		orderStr = multiWrapper.WrapSql(columnsAndFields...)
-	} else {
-		orderStr = columnsAndFields[0]
-	}
-	direction := strings.ToLower(o.direction)
-	switch direction {
-	case "asc", "desc":
-		orderStr += " " + direction
-	case "":
-	default:
-		return "", nil, errors.New(`gorp: Order by direction must be empty string, "asc", or "desc"`)
-	}
-	return orderStr, params, nil
-}
-
-// subQuery is provided to use plan types as sub-queries in from/join
-// clauses.
-type subQuery interface {
-	QuotedTable() string
-	getTable() *gorp.TableMap
-	getTarget() reflect.Value
-	getColMap() structColumnMap
-	errors() []error
-	selectQuery() (string, error)
-	getArgs() []interface{}
-}
 
 // A QueryPlan is a Query.  It returns itself on most method calls;
 // the one exception is Assign(), which returns an AssignQueryPlan (a type of
@@ -198,139 +62,6 @@ type QueryPlan struct {
 	limit          int64
 	offset         int64
 	args           []interface{}
-}
-
-// Query generates a Query for a target model.  The target that is
-// passed in must be a pointer to a struct, and will be used as a
-// reference for query construction.
-func Query(m *gorp.DbMap, exec gorp.SqlExecutor, target interface{}) interfaces.Query {
-	// Handle non-standard dialects
-	switch src := m.Dialect.(type) {
-	case gorp.MySQLDialect:
-		m.Dialect = dialects.MySQLDialect{src}
-	case gorp.SqliteDialect:
-		m.Dialect = dialects.SqliteDialect{src}
-	default:
-	}
-	plan := &QueryPlan{
-		dbMap:    m,
-		executor: exec,
-	}
-
-	targetVal := reflect.ValueOf(target)
-	if targetVal.Kind() != reflect.Ptr || targetVal.Elem().Kind() != reflect.Struct {
-		plan.Errors = append(plan.Errors, errors.New("A query target must be a pointer to struct"))
-	}
-	targetTable, err := plan.mapTable(targetVal)
-	if err != nil {
-		plan.Errors = append(plan.Errors, err)
-		return plan
-	}
-	plan.target = targetVal
-	plan.table = targetTable
-	return plan
-}
-
-func (plan *QueryPlan) getTarget() reflect.Value {
-	return plan.target
-}
-
-func (plan *QueryPlan) getColMap() structColumnMap {
-	return plan.colMap
-}
-
-func (plan *QueryPlan) errors() []error {
-	return plan.Errors
-}
-
-func (plan *QueryPlan) getArgs() []interface{} {
-	return plan.args
-}
-
-func (plan *QueryPlan) getTable() *gorp.TableMap {
-	return plan.table
-}
-
-func (plan *QueryPlan) mapSubQuery(q subQuery) *gorp.TableMap {
-	if len(q.errors()) != 0 {
-		plan.Errors = append(plan.Errors, q.errors()...)
-	}
-	query, err := q.selectQuery()
-	if err != nil {
-		plan.Errors = append(plan.Errors, err)
-	}
-	alias := q.QuotedTable()
-	plan.quotedTable = fmt.Sprintf("(%s) as %s", query, alias)
-	for _, m := range q.getColMap() {
-		m.quotedTable = alias
-		plan.colMap = append(plan.colMap, m)
-	}
-	return q.getTable()
-}
-
-func (plan *QueryPlan) mapTable(targetVal reflect.Value) (*gorp.TableMap, error) {
-	if targetVal.Kind() != reflect.Ptr || targetVal.Elem().Kind() != reflect.Struct {
-		return nil, errors.New("gorp: Cannot create query plan - target value must be a pointer to a struct")
-	}
-
-	if subQuery, ok := targetVal.Interface().(subQuery); ok {
-		return plan.mapSubQuery(subQuery), nil
-	}
-
-	targetTable, err := plan.dbMap.TableFor(targetVal.Type().Elem(), false)
-	if err != nil {
-		return nil, err
-	}
-
-	if err = plan.mapColumns(targetTable, targetVal); err != nil {
-		return nil, err
-	}
-	return targetTable, nil
-}
-
-// mapColumns creates a list of field addresses and column maps, to
-// make looking up the column for a field address easier.  Note that
-// it doesn't do any special handling for overridden fields, because
-// passing the address of a field that has been overridden is
-// difficult to do accidentally.
-func (plan *QueryPlan) mapColumns(table *gorp.TableMap, value reflect.Value) (err error) {
-	value = value.Elem()
-	valueType := value.Type()
-	if plan.colMap == nil {
-		plan.colMap = make(structColumnMap, 0, value.NumField())
-	}
-	queryableFields := 0
-	quotedTableName := plan.dbMap.Dialect.QuotedTableForQuery(table.SchemaName, table.TableName)
-	for i := 0; i < value.NumField(); i++ {
-		fieldType := valueType.Field(i)
-		fieldVal := value.Field(i)
-		if fieldType.Anonymous {
-			if fieldVal.Kind() != reflect.Ptr {
-				fieldVal = fieldVal.Addr()
-			} else if fieldVal.IsNil() {
-				// embedded types must be initialized for querying
-				fieldVal.Set(reflect.New(fieldVal.Type().Elem()))
-			}
-			plan.mapColumns(table, fieldVal)
-		} else if fieldType.PkgPath == "" {
-			col := table.ColMap(fieldType.Name)
-			quotedCol := plan.dbMap.Dialect.QuoteField(col.ColumnName)
-			fieldMap := fieldColumnMap{
-				addr:         fieldVal.Addr().Interface(),
-				column:       col,
-				quotedTable:  quotedTableName,
-				quotedColumn: quotedCol,
-			}
-			plan.colMap = append(plan.colMap, fieldMap)
-			if !col.Transient {
-				queryableFields++
-			}
-		}
-	}
-	if queryableFields == 0 {
-		return errors.New("No fields in the target struct are mappable.")
-	}
-	return
 }
 
 // Extend returns an extended query, using extensions for the
